Added isWikiTag parameter to word, entity and nested wikitags
Added pipe character to the relative entry point for wikitag data reader positions.
Cleaned file saved reader positions from calculations, which can be done later.
Made readerPosInherited non const, this might change later to an internal variable for the adding wikitag the data belongs to.
Cleaned up debug output ever so slightly, so formatting notice isFormatStart, isFormatEnd, FormatType display comes in one line with the element and not on an extra line in addWikiT$
Included formatting in caluclation for the actual readerPos offset.